Restore allies information
Due to the merge on PR #15, the information about the allies pokemon
were removed as parte of the conflict resolution.

This commits brings back the information, and simplifies how to
configure a checkbox, as there is no need to use the extra `store`
attribute to keep track of it.

This change should be backward compatible and should mantain the list of
pokemons saved on the checklist.